cambio del id de integer a long

+++ just-travel-it/src/main/java/is2/justtravelit/services/FlightService.java
@@ -12,8 +12,8 @@ public interface FlightService {
 
     FlightDTO addFlight(FlightDTO request);
 
-    FlightDTO deleteFlight(Integer id);
+    FlightDTO deleteFlight(long id);
 
-    FlightDTO getFlightById(Integer id);
+    FlightDTO getFlightById(long id);
     
 }
