+++ just-travel-it/src/main/java/is2/justtravelit/services/HotelService.java
@@ -11,5 +11,5 @@ public interface HotelService {
 
     public HotelDTO addHotel(HotelDTO request);
 
-    HotelDTO getHotelById(Integer id);
+    HotelDTO getHotelById(long id);
 }
